+++ tools/decode-status.py
@@ -192,7 +192,16 @@ a_features = [[
     "USE_SONOFF_SC","USE_SONOFF_RF","USE_SONOFF_L1","USE_EXS_DIMMER",
     "USE_ARDUINO_SLAVE","USE_HIH6","USE_HPMA","USE_TSL2591",
     "USE_DHT12","USE_DS1624","USE_GPS","USE_HOTPLUG",
-    "USE_NRF24","USE_MIBLE","USE_HM10","",
+    "USE_NRF24","USE_MIBLE","USE_HM10","USE_LE01MR",
+    "","","",""
+    ],[
+    "","","","",
+    "","","","",
+    "","","","",
+    "","","","",
+    "","","","",
+    "","","","",
+    "","","","",
     "","","",""
     ]]
 
@@ -227,7 +236,7 @@ else:
         obj = json.load(fp)
 
 def StartDecode():
-    print ("\n*** decode-status.py v20200207 by Theo Arends and Jacek Ziolkowski ***")
+    print ("\n*** decode-status.py v20200210 by Theo Arends and Jacek Ziolkowski ***")
 
 #    print("Decoding\n{}".format(obj))