+++ tasmota/CHANGELOG.md
@@ -11,6 +11,7 @@
 - Add BootCount Reset Time as BCResetTime to ``Status 1``
 - Add ``ZbZNPReceived``and ``ZbZCLReceived`` being published to MQTT when ``SetOption66 1``
 - Add optional Wifi AccessPoint passphrase define WIFI_AP_PASSPHRASE in my_user_config.h (#7690)
+- Add support for FiF LE-01MR energy meter by saper-2 (#7584)
 
 ### 8.1.0.5 20200126